requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==0.5.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==64.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
Pillow==8.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1

cloudinary==1.33.0

